Fix Connect, Replace and disconnect could do its operation on the wrong connection
Previous solution assumes that a single Input can only be connected to
an output once. That's not true as it can be connected multiple times to
the same output as long as the condition is different. Connect, Replace
and disconnect now accepts a connection as argument so there is no doubt
about the connection the operation should be applied to.