Split board up into sub boards
One of the problems with the board si that it requires a lot of memory to make a large board. Lots of memory leads to lots of cache misses. The idea here is to split the board up into smaller sub boards and there are only sub boards for the places on the board that are actually in use. This means that if routing a wire only goes a short distance on a large board, then it will have a much easier to fitting all the space into cache as there are only sub boards for the space it uses